Overhaul math - constexpr all the things - Fix integer promotion warnings for all types -- Use a single templated type per function -- Use explicit casts after arithmetic operations - Wrap cos[f], sin[f], and sqrt[f] - Add vector aliases for all types - Add a few missing operators - Make all matrix math r32 and remove templates - Restrict Normalize and GetOrbitPos to float vectors

+++ LCDHardwareMonitor/src/ili9341.hpp
@@ -376,7 +376,7 @@ ILI9341_DisplayTest(ILI9341State& ili9341)
 		ILI9341_SetPixel(ili9341, { r, b }, Colors16::Gray);
 	}
 
-	v2u16 quarter = ili9341.size / 4;
+	v2u16 quarter = ili9341.size / (u16) 4;
 
 	if (ili9341.rowColSwap)
 	{
